requirements.txt
absl-py==0.11.0
aiohttp==3.7.4
albumentations==0.5.2
aniso8601==9.0.0
anyio==2.2.0
apispec==4.3.0
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==20.1.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
autograd==1.3
Babel==2.9.0
backcall==0.2.0
backports.functools-lru-cache==1.6.1
bleach==3.3.0
blinker==1.4
cached-property==1.5.1
cachetools==4.2.1
cchardet==2.1.7
certifi==2020.12.5
cffi==1.14.5
click==7.1.2
cloudpickle==1.6.0
cryptography==3.4.4
cycler==0.10.0
cytoolz==0.11.0
dask==2021.3.0
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.3
efficientnet-pytorch==0.6.3
entrypoints==0.3
Flask==1.1.2
flask-apispec==0.11.0
Flask-RESTful==0.3.8
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-auth==1.24.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.36.1
idna==2.10
image-classifiers==1.0.0
imageio==2.9.0
importlib-metadata==3.7.0
importlib-resources==5.1.2
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyter-packaging==0.7.12
jupyter-server==1.4.1
jupyterlab==3.0.10
jupyterlab-pygments==0.1.2
jupyterlab-server==2.3.0
jupyterlab-widgets==1.0.0
Keras==2.4.3
Keras-Applications==1.0.7
Keras-Preprocessing==1.1.2
Markdown==3.3.4
MarkupSafe==1.1.1
marshmallow==3.10.0
mistune==0.8.4
multidict==5.1.0
munch==2.5.0
nbclassic==0.2.6
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.4.3
networkx==2.5
notebook==6.2.0
numpy==1.20.1
oauthlib==3.0.1
olefile==0.46
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.2
parso==0.8.1
patsy==0.5.1
PennyLane==0.14.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
pip==21.0.1
pooch==1.3.0
pretrainedmodels==0.7.4
prometheus-client==0.9.0
promise==2.3
prompt-toolkit==3.0.16
protobuf==3.15.5
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.20
Pygments==2.8.0
PyJWT==2.0.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.0
seaborn==0.11.1
segmentation-models-pytorch==0.1.3
semantic-version==2.6.0
Send2Trash==1.5.0
setuptools==54.0.0
Shapely==1.5.9
six==1.15.0
sniffio==1.2.0
statsmodels==0.12.2
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2021.3.17
timm==0.3.2
toml==0.10.2
toolz==0.11.1
torch==1.8.0
torchvision==0.9.0
tornado==6.1
tqdm==4.58.0
traitlets==5.0.5
typeguard==2.11.1
typing-extensions==3.7.4.3
urllib3==1.26.3
wcwidth==0.2.5
webargs==7.0.1
webencodings==0.5.1
Werkzeug==1.0.1
wheel==0.36.2
widgetsnbextension==3.5.1
wrapt==1.12.1
yarl==1.6.3
zipp==3.4.0
